Userspace vhost: memory file mapping
This patch fixes an issue seen in Ubuntu where the return value of the
function realpath() is not dealt with. This return value along with the
resolved path are now checked. Previously if no path was found it would
result in an error. Now it continues to the next path which is the
desired behaviour. If, after this, all paths are checked and the memory file
isn't found then an error is the result.

Change-Id: I17c0e7af4051174eac35b50c1f5624af76b28987
Reviewed-by: Mark D. Gray <mark.d.gray@intel.com>
Signed-off-by: Anthony Fee <anthonyx.fee@intel.com>