docs: Add plaintext Getting Started Guide
Previously the documentation for Intel(R) DPDK vSwitch was published
as a PDF document. This was difficult for the community to edit.
Additionally the document had become cluttered and fragmented as more
features were added to DPDK vSwitch.

This commit resolves this issue by coverting the documentation to
plaintext - specifically GitHub Flavored Markdown - and moving it
into the mainline source.

Change-Id: Ic5a032725116b7c06b156ca443776ed040938b5c
Signed-off-by: Stephen Finucane <stephen.finucane@intel.com>
Signed-off-by: Mark B. Kavanagh <mark.b.kavanagh@intel.com>
Reviewed-by: Mark D. Gray <mark.d.gray@intel.com>
Reviewed-by: Eimear Morrissey <eimear.morrissey@intel.com>
Reviewed-by: Meghan Halton <meghan.halton@intel.com>