flow leak and multiple flow entry variables with same name (2nd submittal)
Fixes memory leak in switch subroutine. For packets going through the switch subroutine switch_packet(), not all paths result in the packet being processed, and since the calling code assumes that the packet is handled, packet buffers and packets are lost causing a memory leak.
The variable "used" used twice in the flow entry of the base structure and statistics structure, however is used in three separate variables in the flow code file:
1): In the flow stats structure, boolean value 0 value to represent a flow entry not used yet case.
2): In the flow stats structure, 64 bit count for the CPU clock value.
3): In the flow base structure to represent a valid flow entry.
4): In the vswitchd structure to pass time last used to the vswitchd.
The "used" variable is used too many times and creates an illogical usage pattern when assuming the variable is the same across the structure, requiring extra work to determine its exact usage. Renamed the variable for case 3 to better represent it's logical usage.

Signed-off-by: Mike A. Polehn <mike.a.polehn@intel.com>