Userspace Vhost: Resolve issue which causes QEMU to fall back on userspace virtio
Patch to correct userspace vhost issue introduced by previous patch to fix
compilation failure upon warning of unused results returned by
realpath()

A previous commit removed a condition from an if statement and consequentialy
introduced an issue whereby the aforementioned if statement always executes.
A break inside this if statement results in the instructions necessary for
successful userspace vhost initialisation are not executed. The patch resolves
this by removing the if statement and associated break.

Change-Id: I5ff8cec0d28b750fae9fb489677deeb4db4c6b41
Reviewed-by: Mark D. Gray <mark.d.gray@intel.com>
Signed-off-by: cloftus <ciara.loftus@intel.com>