ofproto: Increase ovs flow entry timeout from 5s to 30s
The default kernel-space flow entry timeout of 5 second is too fast
for running RFC2544 throughput test using Spirent. Spirent can be
configured to send a verification stream to establish flows in the
flow table. However, the time between the end of verification stream
and the start of actual test stream is greater than 5 second which is not
 configurable on the RFC2544 test setup. Hence, under a zero loss
condition exception packets are lost for every stream of flows
resulting in very low rates. Increasing the flow entry timeout from
 5 second to 30 second persists the flows in the flow table long
enough sustain good performance.
Declared a constant KERNEL_FLOW_TABLE_TIMEOUT and set to 30000.

Change-Id: Ie17fd18b4c8fd6e9f1ae1154e15d870b4622daa9
Reviewed-by: Mark D. Gray <mark.d.gray@intel.com>
Signed-off-by: Rashmi Pujar <rashmi.c.pujar@intel.com>