vswitch: Fix recv_upcall error for vhost-virtio use-case
dpif-dpdk does not translate OVS_ACTION_ATTR_USERSPACE
actions with USER_ACTION_COOKIE_SLOW_PATH into datapath
actions.

When some packets (for example, IPv6 ND protocol) are sent
to vswitchd as miss upcalls, the vswitchd attempts to put
OVS_ACTION_ATTR_USERSPACE actions in the datapath. The datapath
then receives a put message with no actions. This causes
subsequent packets in this flow to be sent to the
exception ring without upcall and be incorrectly interpreted
by dpif_recv().

This patch adds partial support for the OVS_ACTION_ATTR_USERSPACE
action in order to resolve this issue. It also adds a check
that no flow messages with num_actions == 0 are sent to datapath.

Change-Id: I5925bf25a8f8983d5bb8a6b6e6f0e44b2e4fc1a7
Reviewed-by: Ciara Loftus <ciara.loftus@intel.com>
Reviewed-by: Ian Stokes <ian.stokes@intel.com>
Reviewed-by: Mark B. Kavanagh <mark.b.kavanagh@intel.com>
Signed-off-by: Mark D. Gray <mark.d.gray@intel.com>