ovdk-flow: Add prefetching to the flow extraction
Each time a packet is received by an rte_port, the flow
key is extracted from the packet. To extract the flow key,
two pieces of memory are needed.

1. The mbuf metadata where the key and the signature will
be added to
2. The packet data where the key will be extracted from.

In order to maximise the probability that this memory
is in the cache, this memory should be prefetched.

Change-Id: I6c1266844d18b37e24cc982ff9a207b8810f66ec
Reviewed-by: Mark B. Kavanagh <mark.b.kavanagh@intel.com>