datapath: Implement packet framework userspace vhost ports
The packet framework provides an abstract interface
'rte_port' that can be implemented in order to interface
an 'rte_pipeline' with a network device. This network device
can be a physical or virtual device.

This patch introduces a new implementation of 'rte_port'
for user-space vhost devices. This patch also enables these
ports to be added to an 'ovdk_pipeline'.

This patch initializes all the resources required for the
ports at start up. It also populates structures required
by 'ovdk_pipeline'

Change-Id: I76f4f5a77ba0bd67ae82a0e5f49c9ceec6934742
Signed-off-by: Mark D. Gray <mark.d.gray@intel.com>
Signed-off-by: Kevin Traynor <kevin.traynor@intel.com>
Signed-off-by: Eimear Morrissey <eimear.morrissey@intel.com>
Signed-off-by: Meghan Halton <meghan.halton@intel.com>
Signed-off-by: Mark B. Kavanagh <mark.b.kavanagh@intel.com>
Signed-off-by: Stephen Finucane <stephen.finucane@intel.com>
Signed-off-by: Ian Stokes <ian.stokes@intel.com>
Signed-off-by: Ciara Loftus <ciara.loftus@intel.com>