datapath: Introducing 'ovdk_pipeline'
'ovdk_pipeline' implements an adapter for 'rte_pipeline'. 'rte_pipeline' is
an object in the DPDK packet framework:

    http://www.dpdk.org/ml/archives/dev/2014-June/003059.html

Each 'rte_pipeline' is wrapped by an 'ovdk_pipeline'. We use one
'rte_pipeline' per core in order to process packets for that core.

ovdk_pipeline_run() is routinely called in order to
process packets received by that pipeline.

The use of 'rte_pipeline' will:

1. Allow the vswitch to quickly take advantage of features developed
by DPDK for the packet framework

2. Allow easier customization of the dataplane

3. Enable hardware offload of some functionality to improve performance

Change-Id: I1b303eb466f7c187262c365df8a3d65844420caa
Signed-off-by: Mark D. Gray <mark.d.gray@intel.com>
Signed-off-by: Kevin Traynor <kevin.traynor@intel.com>
Signed-off-by: Eimear Morrissey <eimear.morrissey@intel.com>
Signed-off-by: Meghan Halton <meghan.halton@intel.com>
Signed-off-by: Mark B. Kavanagh <mark.b.kavanagh@intel.com>
Signed-off-by: Stephen Finucane <stephen.finucane@intel.com>
Signed-off-by: Ian Stokes <ian.stokes@intel.com>
Signed-off-by: Ciara Loftus <ciara.loftus@intel.com>