datapath: Add infrastructure for different port types
Each port associated with the datapath is called a vport and
has an individual unique index called the 'vportid'. Currently,
this is equivalent to the concept of an Open vSwitch odp port
number.

Unlike previous versions of the vswitch, the ranges of these
numbers can be arbitrarily defined at compile time.
'ovdk_vport_types.h' defines these ranges and as such, needs to
be shared with the vswitchd.

At initialization time, resources for all vports must be
allocated. Then, as the vswitchd adds ports these vports
get added to the rte_pipeline. Each vport is described
in a memzone which can be shared to a guest if required. The
data in this memzone is described in 'ovdk_vport_info.h' and
can be accessed by ovs_vport.

If a port works with the DPDK packet framework, it should be
reasonably easy to enable this port with the vswitch.

Change-Id: Ie45378238cc401343ac850d05f9b8eb13e87adda
Signed-off-by: Mark D. Gray <mark.d.gray@intel.com>
Signed-off-by: Kevin Traynor <kevin.traynor@intel.com>
Signed-off-by: Eimear Morrissey <eimear.morrissey@intel.com>
Signed-off-by: Meghan Halton <meghan.halton@intel.com>
Signed-off-by: Mark B. Kavanagh <mark.b.kavanagh@intel.com>
Signed-off-by: Stephen Finucane <stephen.finucane@intel.com>
Signed-off-by: Ian Stokes <ian.stokes@intel.com>
Signed-off-by: Ciara Loftus <ciara.loftus@intel.com>