datapath: Implement packet framework ethdev ports
The packet framework provides an abstract interface
'rte_port' that can be implemented in order to interface
an 'rte_pipeline' with a network device. This network device
can be a physical or virtual device.

An implementation of this interface for ethdev ports
is provided in DPDK. This patch allows for these ethdev
ports to be added to an 'ovdk_pipeline'. It is mainly
concerned with initialization of the ethernet devices
and population of structures required by 'ovdk_pipeline'

Change-Id: I801e60fb06ee3c76d39838648dd28613a284b345
Signed-off-by: Mark D. Gray <mark.d.gray@intel.com>
Signed-off-by: Kevin Traynor <kevin.traynor@intel.com>
Signed-off-by: Eimear Morrissey <eimear.morrissey@intel.com>
Signed-off-by: Meghan Halton <meghan.halton@intel.com>
Signed-off-by: Mark B. Kavanagh <mark.b.kavanagh@intel.com>
Signed-off-by: Stephen Finucane <stephen.finucane@intel.com>
Signed-off-by: Ian Stokes <ian.stokes@intel.com>
Signed-off-by: Ciara Loftus <ciara.loftus@intel.com>