datapath: Implement packet framework bridge ports
Packet framework provides an abstract interface
'rte_port' that can be implemented in order to interface
an 'rte_pipeline' with a network device. This network device
can be a physical or virtual device.

An implementation of this interface for sink and source ports
is provided in DPDK.

This patch adds these sink/source ports as bridge ports to an
'ovdk_pipeline'. It is mainly concerned with population of
structures required by 'ovdk_pipeline'.

In standard Open vSwitch, bridge ports provide a standard
network interface on the host for each bridge port (e.g.
br0, br1). This interface works like any normal Linux
interface (i.e. can be viewed with ifconfig, can accept
'ping' requests, etc). This functionality is not available
with this implementation of a bridge port.

Change-Id: Iabf0dcc36409a18b85f33aab9ad0984f95ced1e3
Signed-off-by: Mark D. Gray <mark.d.gray@intel.com>
Signed-off-by: Kevin Traynor <kevin.traynor@intel.com>
Signed-off-by: Eimear Morrissey <eimear.morrissey@intel.com>
Signed-off-by: Meghan Halton <meghan.halton@intel.com>
Signed-off-by: Mark B. Kavanagh <mark.b.kavanagh@intel.com>
Signed-off-by: Stephen Finucane <stephen.finucane@intel.com>
Signed-off-by: Ian Stokes <ian.stokes@intel.com>
Signed-off-by: Ciara Loftus <ciara.loftus@intel.com>