datapath: Introducing 'ovs-dpdk'
'ovs-dpdk' is a user space DPDK application that implements
a fast dataplane using the Packet Framework. This dataplane
can be controlled using control messages sent via DPDK rings.

Each lcore runs an instantiation of an 'ovdk_datapath' and an
'ovdk_pipeline'.

Unlike previous versions of the vswitch, this implementation
allows the datapath to scale down to one DPDK lcore or up
to an arbitrary number of lcores by specifying a core mask
when starting the 'ovs-dpdk' application.

Work is distributed between lcores at the granularity of an
input port. Therefore, if you are using two ports and three
cores, one core is effectively not in use. Currently, when
adding ports to the switch, they are added to cores in a
round robin manner. This is different to the previous versions
of the vswitch as one core was dedicated to switching
virtual machine ports and 2 were dedicated to physical ports.
This created a bottleneck for some use cases.

Stats can be defined to start on an arbitrary port defined
at the command line.

Change-Id: I3e6fba9f9da96d7b8d17677c7545176ca39d9db0
Signed-off-by: Mark D. Gray <mark.d.gray@intel.com>
Signed-off-by: Kevin Traynor <kevin.traynor@intel.com>
Signed-off-by: Eimear Morrissey <eimear.morrissey@intel.com>
Signed-off-by: Meghan Halton <meghan.halton@intel.com>
Signed-off-by: Mark B. Kavanagh <mark.b.kavanagh@intel.com>
Signed-off-by: Stephen Finucane <stephen.finucane@intel.com>
Signed-off-by: Ian Stokes <ian.stokes@intel.com>
Signed-off-by: Ciara Loftus <ciara.loftus@intel.com>