dpif-dpdk: Introducing 'dpif-dpdk-flow-table'
'dpif-dpdk-flow-table.c' implements a flow table in
the vswitchd used to keep track of flows in the datapath.
Currently, this uses a simple array table. Future work
should modify this to improve look up performance.

As 'ovs-dpctl' does not use the vswitchd the state of this
table needs to be stored in a memory mapped file
so that it can be persistent across invocations of
'ovs-dpctl'.

Change-Id: I419f9c42ea45440022a35d7b6b9f5ae687a976d4
Signed-off-by: Mark D. Gray <mark.d.gray@intel.com>
Signed-off-by: Kevin Traynor <kevin.traynor@intel.com>
Signed-off-by: Eimear Morrissey <eimear.morrissey@intel.com>
Signed-off-by: Meghan Halton <meghan.halton@intel.com>
Signed-off-by: Mark B. Kavanagh <mark.b.kavanagh@intel.com>
Signed-off-by: Stephen Finucane <stephen.finucane@intel.com>
Signed-off-by: Ian Stokes <ian.stokes@intel.com>
Signed-off-by: Ciara Loftus <ciara.loftus@intel.com>