dpif-dpdk: Introducing 'dpif-dpdk-vport-table'
'dpif-dpdk-vport-table.c' implements a table in
the vswitchd used to keep track of vports in the datapath.
In particular, it needs to track to which lcore a
vport has been added to.

Currently, this uses a simple array table. Future work
should modify this to improve look up performance.

As 'ovs-dpctl' does not use the vswitchd the state of this
table needs to be stored in a memory mapped file
so that it can be persistent across invocations of
'ovs-dpctl'.

Change-Id: Ia8d34626bfe345d96690571d0296caddf8037799
Signed-off-by: Mark D. Gray <mark.d.gray@intel.com>
Signed-off-by: Kevin Traynor <kevin.traynor@intel.com>
Signed-off-by: Eimear Morrissey <eimear.morrissey@intel.com>
Signed-off-by: Meghan Halton <meghan.halton@intel.com>
Signed-off-by: Mark B. Kavanagh <mark.b.kavanagh@intel.com>
Signed-off-by: Stephen Finucane <stephen.finucane@intel.com>
Signed-off-by: Ian Stokes <ian.stokes@intel.com>
Signed-off-by: Ciara Loftus <ciara.loftus@intel.com>