vswitchd: Modify 'dpif-dpdk' to use packet framework enabled datapath
There is no easy way to do this and maintain a bisectable
code base as this change touches most aspects of the 'dpif-dpdk'
implementation. As a result, this patch is large but maintains
a bisectable history.

This removes the non packet framework datapath and
replaces it with the modified datapath using packet framework.
It also modifies 'dpif-dpdk' to use this modified datapath. In
summary, this patch enables 'ovs-vswitchd' to interface with
'ovs-dpdk'.

The main changes made to enable this are:

- Modify 'dpif-dpdk' to use the updated message structures for
communication with 'ovs-dpdk'
- Use the 'dpif-dpdk-vport-table' and 'dpif-dpdk-flow-table' to
track flows and vports added to ovs-dpdk
- Add logic to detect the number of lcores in the datapath
- Add code to distribute vports between lcores. A vport is
added as an in-port to one lcore but must be added as an out-port
to every core. This is because any lcore can output to any port.
- Update 'ovs-dpctl' to work with 'ovs-dpdk'
- Update the 'ovs_client' guest application
- Modify code to compile against DPDK 1.7.0
- Update all the unit tests to work with packet framework
enabled datapath.
- Remove any unit tests that are no longer relevant

Change-Id: I53b203113263de6d3bd765972641bea46ee62308
Signed-off-by: Mark D. Gray <mark.d.gray@intel.com>
Signed-off-by: Kevin Traynor <kevin.traynor@intel.com>
Signed-off-by: Eimear Morrissey <eimear.morrissey@intel.com>
Signed-off-by: Meghan Halton <meghan.halton@intel.com>
Signed-off-by: Mark B. Kavanagh <mark.b.kavanagh@intel.com>
Signed-off-by: Stephen Finucane <stephen.finucane@intel.com>
Signed-off-by: Ian Stokes <ian.stokes@intel.com>
Signed-off-by: Ciara Loftus <ciara.loftus@intel.com>