bugfix: Workaround for duplicated packets issue
When running with Libvirt, duplicated packets were observed when
VMs were hard rebooted and brought up again. This was due to an issue
with port deletion in Intel DPDK vSwitch.

This temporary workaround disables deletion such that ports which
have been used cannot be used again. A more permanent solution for
this problem is in the works.

Change-Id: Iee702c4b20e34eedd1902eb28327f96c27ec4e22
Signed-off-by: Meghan Halton <meghan.halton@intel.com>
Reviewed-by: Kevin Traynor <kevin.traynor@intel.com>
Reviewed-by: Mark D. Gray <mark.d.gray@intel.com>