dpif: fix dpif_dpdk_get_stats
'dpif_dpdk_get_stats' requests statistics from every pipeline with
which the relevant vportid is associated with (i.e. the in_port and
all out_ports). Stats received from each pipeline are summated and
returned to the dpif. However, at the datapath layer, each reply
sent to the dpif already contains the total summated stats.

Resolve this issue by removing the duplication of stats summation
and reduce the number of transactions between the dpif and the
datapath to just one, between the dpif and the pipleine on which
the in_port resides.

Change-Id: Ia751749ad22d4fca8cda27d02b88ed229d0cffb2
Signed-off-by: Mark Kavanagh <mark.b.kavanagh@intel.com>
Reviewed-By: Stephen Finucane <stephen.finucane@intel.com>
Reviewed-By: Mark Gray <mark.d.gray@intel.com>