dpif-dpdk: Implement new datapath port state
Firstly, when adding a port, check if the port has been previously
added, by checking the lcoreid value associated with the port. If set
to a valid value, the port exists and merely needs to be re-enabled.
If not, the port does not exist and must be added and enabled.

Secondly, when deleting a port, ensure that both in-ports *and*
out-ports are deleted.

Change-Id: I8c0982b2f10b11ae9f408aa4db08c973ffbe23bf
Signed-off-by: Ian Stokes <ian.stokes@intel.com>
Signed-off-by: Mark B. Kavanagh <mark.b.kavanagh@intel.com>
Signed-off-by: Stephen Finucane <stephen.finucane@intel.com>
Reviewed-by: Kevin Traynor <kevin.traynor@intel.com>
Reviewed-by: Stephen Finucane <stephen.finucane@intel.com>
Reviewed-by: Ciara Loftus <ciara.loftus@intel.com>