Set port_in_id correctly to re-enable ports
The port_in_id variable is set when a pipeline in_port is
created. However when a port re-enable is required, the
in_port has already been created and the port_in_id already
set in the vport_info entry for that port. Port_in_id is
initialised to 0 and then used as the port_in_id to enable a
port previously created. This fix adds a call to get the
port_in_id from the vport_info entry so that the correct
port_in_id can be called with rte_in_port_enable.

*ovdk_pipeline: added call to ovdk_vport_get_in_portid for
 ports in the disabled state.

Change-Id: I53a0734828a8de343dfe491129f24f229b7fac73
Signed-off-by: Ian Stokes <ian.stokes@intel.com>
Reviewed-by: Kevin Traynor <kevin.traynor@intel.com>
Reviewed-by: Stephen Finucane <stephen.finucane@intel.com>