dpif-dpdk: Fix bug with 'flow_get'
The 'ovs-vswitchd' application segfaults under certain conditions, such
as when running certain OFTest tests. This is as a result of
"revalidator" changes introduced between Open vSwitch v2.0 and v2.1.2.
Resolve this issue by creating a required 'ofpbuf'. There is no need to
free this mbuf as this is handled by the caller.

Change-Id: I58a7eff4d462567406a655f92f5b8bdbc1035afd
Signed-off-by: Stephen Finucane <stephen.finucane@intel.com>
Reviewed-by: Kevin Traynor <kevin.traynor@intel.com>
Reviewed-by: Mark D. Gray <mark.d.gray@intel.com>